virito-blk: Support NOSPC error
Add a status code to indicate that the storage is full on the virtio-blk
device side.

This is useful in scenarios where a virtio-blk disk image is provided on a
sparse disk and the host storage is full.
In this scenario, the driver cannot create a new file to the disk image
because the sparse disk cannot be expanded.
However, if the host have a service that clean up host stroage regularly
by wiping cache files, the guest may want to retry the request.
For this type of special handling, this error should have a separate
error code in virito-blk.

Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>